chore: move `take_until_unbalanced` to vm/serde
Since it's a single function, it's better to simplify deployment by
avoiding an extra crate.
