+++ .github/workflows/publish.yml
@@ -14,10 +14,6 @@ jobs:
       uses: actions/checkout@v2
     - name: Install stable toolchain
       uses: dtolnay/rust-toolchain@1.69.0
-    - name: Publish crate cairo-take_until_unbalanced
-      env:
-          CRATES_TOKEN: ${{ secrets.CARGO_REGISTRY_TOKEN }}
-      run: cargo publish --token ${CRATES_TOKEN} --all-features --manifest-path ./deps/parse-hyperlinks/Cargo.toml
     - name: Publish crate cairo-felt
       env:
           CRATES_TOKEN: ${{ secrets.CARGO_REGISTRY_TOKEN }}
