+++ Cargo.toml
@@ -4,7 +4,6 @@ members = [
     "felt",
     "vm",
     "hint_accountant",
-    "./deps/parse-hyperlinks",
 ]
 exclude = ["ensure-no_std"]
 
@@ -47,14 +46,6 @@ anyhow = { version = "1.0.69", default-features = false }
 thiserror = { version = "1.0.32", default-features = false }
 thiserror-no-std = { version = "2.0.2", default-features = false }
 
-# This crate has only one function `take_until_unbalanced` that is
-# very useful for our parsing purposes:
-# https://stackoverflow.com/questions/70630556/parse-allowing-nested-parentheses-in-nom
-# There is a proposal for extending nom::delimited to use this function:
-# https://github.com/Geal/nom/issues/1253
-parse-hyperlinks = { package = "cairo-take_until_unbalanced", path = "./deps/parse-hyperlinks", version = "0.32.0", default-features = false, features = [
-    "alloc",
-] }
 felt = { package = "cairo-felt", path = "./felt", version = "0.8.0", default-features = false, features = [
     "alloc",
 ] }
