+++ RELEASE.md
@@ -17,8 +17,6 @@
     - `cairo-vm-cli/Cargo.toml`: update the version string and also the
       `cairo-vm` dependency version to match the above.
     - `felt/Cargo.toml`: update the version string.
-    - `deps/parse-hyperlinks/Cargo.toml`: this vendored dependency needs its
-      version bumped, but does not need to match the other crate versions.
   - [Here](https://github.com/lambdaclass/cairo-rs/pull/1257/files) is an
     example pull request with these changes.
 - [ ] Run `cargo update` and `git add Cargo.lock`
@@ -36,7 +34,6 @@
       versions.
   - [cairo-vm](https://crates.io/crates/cairo-vm)
   - [cairo-felt](https://crates.io/crates/cairo-felt)
-  - [cairo-take-until-unbalanced](https://crates.io/crates/cairo-take_until_unbalanced)
 - [ ] Create a release in Github.
   - Select the recently created tag.
   - Set the title to the version string.
