+++ vm/Cargo.toml
@@ -16,7 +16,6 @@ std = [
     "bincode/std",
     "anyhow/std",
     "starknet-crypto/std",
-    "parse-hyperlinks/std",
     "felt/std",
     "dep:num-prime",
 ]
@@ -59,13 +58,6 @@ thiserror-no-std = { workspace = true }
 
 # only for std
 num-prime = { version = "0.4.3", features = ["big-int"], optional = true }
-
-# This crate has only one function `take_until_unbalanced` that is
-# very useful for our parsing purposes:
-# https://stackoverflow.com/questions/70630556/parse-allowing-nested-parentheses-in-nom
-# There is a proposal for extending nom::delimited to use this function:
-# https://github.com/Geal/nom/issues/1253
-parse-hyperlinks = { workspace = true }
 felt = { workspace = true }
 bitvec = { workspace = true }
